Ticketing/orders/src/events/listener/ticket-updated-listener.ts
import { Listener } from '@ferhatadibelli/ferhatadibelli/build/events/base-listener';
import { Subjects } from '@ferhatadibelli/ferhatadibelli/build/events/subjects';
import { TicketUpdatedEvent } from '@ferhatadibelli/ferhatadibelli/build/events/ticket-updated-event';
import { Message } from 'node-nats-streaming';
import { Ticket } from '../../models/ticket';
import { queueGroupName } from '../queue-group-name';

export class TicketUpdatedListener extends Listener<TicketUpdatedEvent> {
  readonly subject = Subjects.TicketUpdated;
  queueGroupName = queueGroupName;
  async onMessage(data: TicketUpdatedEvent['data'], msg: Message) {
    const { price, title } = data;
    const ticket = await Ticket.findByEvent(data);

    if (!ticket) {
      throw new Error('Ticket is not found');
    }

    ticket.set({ price, title });
    await ticket.save();

    msg.ack();
  }
}
