Ticketing/orders/src/events/listener/__test__/ticket-updated-listener.test.ts
import { TicketUpdatedEvent } from '@ferhatadibelli/ferhatadibelli/build/events/ticket-updated-event';
import mongoose from 'mongoose';
import { Message } from 'node-nats-streaming';
import { Ticket } from '../../../models/ticket';
import { natsWrapper } from '../../../nats-wrapper';
import { TicketUpdatedListener } from '../ticket-updated-listener';

const setUp = async () => {
  const listener = new TicketUpdatedListener(natsWrapper.client);

  const ticket = Ticket.build({
    id: new mongoose.Types.ObjectId().toHexString(),
    title: 'Concert',
    price: 2323,
  });

  await ticket.save();

  const data: TicketUpdatedEvent['data'] = {
    id: ticket.id,
    title: 'Concert New Data',
    price: 3500,
    userId: 'userId12345',
    version: ticket.version + 1,
  };

  //@ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, data, msg, ticket };
};

it('finds, updates and save an ticket', async () => {
  const { listener, data, msg, ticket } = await setUp();

  await listener.onMessage(data, msg);
  const updatedTicket = await Ticket.findById(ticket.id);
  expect(updatedTicket!.version).toEqual(data.version);
  expect(updatedTicket!.title).toEqual(data.title);
  expect(updatedTicket!.price).toEqual(data.price);
});

it('acks the message', async () => {
  const { listener, data, msg } = await setUp();

  await listener.onMessage(data, msg);
  expect(msg.ack).toHaveBeenCalled();
});

it('doesnt trigger ack function because of skipped version', async () => {
  const { listener, data, msg } = await setUp();

  data.version = 10;

  try {
    await listener.onMessage(data, msg);
  } catch (error) {}

  expect(msg.ack).not.toHaveBeenCalled();
});
